requirements.txt
Flask
pyalsaaudio